@@ -5,32 +5,31 @@ homepage: https://github.com/Workiva/test_html_builder
 description: >
   Builder that generates HTML files from templates for dart browser tests.
 
-
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 executables:
   browser_aggregate_tests:
 
 dependencies:
-  args: ^2.0.0
-  build: ^2.0.0
-  dart_style: ^2.0.0
-  glob: ^2.0.0
-  json_annotation: ^4.0.0
-  path: ^1.6.4
-  test: ^1.15.7
-  test_core: ^0.3.11+4
-  yaml: ^3.0.0
+  args: ^2.1.1
+  build: ^2.0.3
+  dart_style: ^2.0.2
+  glob: ^2.0.1
+  json_annotation: ^4.1.0
+  path: ^1.8.0
+  test: ^1.17.10
+  test_core: ^0.4.0
+  yaml: ^3.1.0
 
 dev_dependencies:
   build_runner: ^2.0.6
   build_test: ^2.1.2
   build_web_compilers: ^3.0.0
-  dependency_validator: ^3.0.0
-  pedantic: ^1.8.0
+  dependency_validator: ^3.1.0
+  pedantic: ^1.11.1
   test_descriptor: ^2.0.0
-  test_process: ^2.0.0
+  test_process: ^2.0.2
 
   # If changes are made to `lib/src/config.dart` and regeneration is needed,
   # uncomment this and comment out the test_html_builder definition in
