@@ -74,7 +74,7 @@ class AggregateTestBuilder extends Builder {
   @override
   Future<void> build(BuildStep buildStep) async {
     _config ??= await decodeConfig(buildStep);
-    if (!_config.browserAggregation) {
+    if (!_config!.browserAggregation) {
       log.fine('browser aggregation disabled');
       return;
     }
@@ -83,21 +83,21 @@ class AggregateTestBuilder extends Builder {
     final isDefault = templatePath == 'test/templates/default_template.html';
     final testGlobs = isDefault
         ? [Glob('test/**_test.dart')]
-        : _config.templateGlobs[templatePath] ?? [];
+        : _config!.templateGlobs[templatePath] ?? [];
     log.fine(
         'Test globs found for template: ${buildStep.inputId}:\n${testGlobs.join('\n')}');
 
     final higherPrecedenceGlobs = <Glob>[];
     if (isDefault) {
       // For the default template, all defined globs are higher precedence.
-      for (final globs in _config.templateGlobs.values) {
+      for (final globs in _config!.templateGlobs.values) {
         higherPrecedenceGlobs.addAll(globs);
       }
     } else {
-      for (final t in _config.templateGlobs.keys) {
+      for (final t in _config!.templateGlobs.keys) {
         // Only templates defined before the current one take precedence.
         if (t == templatePath) break;
-        higherPrecedenceGlobs.addAll(_config.templateGlobs[t]);
+        higherPrecedenceGlobs.addAll(_config!.templateGlobs[t]!);
       }
     }
 
@@ -167,7 +167,7 @@ $mains}
   bool _isBrowserTest(Metadata testMetadata) => _browserRuntimes
       .any((r) => testMetadata.testOn.evaluate(SuitePlatform(r)));
 
-  TestHtmlBuilderConfig _config;
+  TestHtmlBuilderConfig? _config;
 }
 
 /// Builder that uses templates to generate HTML files for dart tests.
@@ -188,7 +188,7 @@ class TemplateBuilder implements Builder {
 
   static AssetId getHtmlId(AssetId assetId) => assetId.changeExtension('.html');
 
-  AssetId getTemplateId(
+  AssetId? getTemplateId(
       Map<String, Iterable<Glob>> templates, AssetId assetId) {
     if (assetId.path.endsWith('.browser_aggregate_test.dart')) {
       return AssetId(assetId.package,
@@ -196,7 +196,7 @@ class TemplateBuilder implements Builder {
     }
 
     for (final templatePath in templates.keys) {
-      final globs = templates[templatePath];
+      final globs = templates[templatePath]!;
       if (globs.any((glob) => glob.matches(assetId.path))) {
         return AssetId(assetId.package, templatePath);
       }
@@ -217,7 +217,7 @@ class TemplateBuilder implements Builder {
     }
 
     _config ??= await decodeConfig(buildStep);
-    final templateId = getTemplateId(_config.templateGlobs, buildStep.inputId);
+    final templateId = getTemplateId(_config!.templateGlobs, buildStep.inputId);
     if (templateId == null) {
       return;
     }
@@ -244,7 +244,7 @@ class TemplateBuilder implements Builder {
     await buildStep.writeAsString(htmlId, htmlContents);
   }
 
-  TestHtmlBuilderConfig _config;
+  TestHtmlBuilderConfig? _config;
 }
 
 class DartTestYamlBuilder extends Builder {
